Fix loading of mask as uint8

+++ tf_semantic_segmentation/datasets/dataset.py
@@ -129,7 +129,8 @@ class Dataset(object):
         @tf.function
         def load(image_path: str, mask_path: str, difficulty: int = None):
             image = load_image(image_path, tf.float32, squeeze=False, channels=channels)
-            mask = load_image(mask_path, tf.int64, squeeze=True, channels=1)
+            mask = load_image(mask_path, tf.uint8, squeeze=True, channels=1)
+
             if difficulty != None:
                 return image, mask, difficulty
             else:
